This branch contains the code written in Processing.
Processing is a Java based coding platform that allowed real-time GUI and backed computation in real-time. Personally I find it easier to program than Java programming platform such as Eclipse and not as memory consuming as the Matlab. Therefore it’s chosen for the project.

For more information: https://processing.org/

P.S: since all of the code is written in Processing, the filename might not make sense as some of them are only using for practice purpose only. These files, however, are backed up here for future reference and usage.